adding details to the extraction table
