+++ src/App.css
@@ -38,8 +38,13 @@ a {
 	left: 0;
 	width: 17vw;
 	height: 100vh;
-	background-image: url("./assets/banner-bg.png");
-	background-color: #08762e;
+	background-image: url("./assets/background-clr-2.jpg");
+	background-repeat: no-repeat;
+	background-attachment: fixed;
+	background-position: center;
+	background-size: cover;
+	background: linear-gradient(rgba(8, 118, 46, 0.3), rgba(8, 118, 46, 0.7)), url("./assets/background-clr-2.jpg");
+	/* background-color: #08762e; */
 	opacity: 1;
 	transform: translateX(0);
 	transition: all .8s ease;
@@ -67,12 +72,37 @@ a {
 	border-left: 5px solid transparent;
 }
 
+.nav_items .dropdown div {
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	gap: 5px;
+	padding: 20px 30px;
+	font-size: 14px;
+	font-weight: 400;
+	text-align: left;
+	transition: all .4s ease;
+	border-left: 5px solid transparent;
+}
+
 .nav_items a:hover,
 .nav_items a.active {
 	font-weight: bold;
 	background: #e6e7ea;
 	border-left: 7px solid #08762e;
+	border-top: .3px solid #08762e;
 	color: #08762e;
+	cursor: pointer;
+}
+.nav_items .notActive {
+	font-weight: bold;
+	background: #e6e7ea;
+	border-top: 0.1px solid #08762e; 
+	border-left: 15px solid #08762e;
+	color: #08762e;
+}
+.nav_items .notActive:hover{
+	cursor: pointer;
 }
 
 .content {
@@ -238,7 +268,8 @@ table {
 	border-radius: 10px;
 	border-radius: 5px;
 	overflow: auto;
-	color: rgb(90, 90, 90);;
+	color: rgb(90, 90, 90);
+	font-size: 13px;
 }
 table thead th {
 	padding: 10px 10px;
@@ -375,29 +406,31 @@ table .res-head {
 	width: 50%;
 }
 .data .form-content form .names div label{
-	font-weight: 700;
+	font-weight: 500;
 	color: #9c9c9c;
+	font-size: 13px;
 }
 .data .form-content form .names input{
-	border: 3px solid rgb(83, 83, 86);
+	border: 1px solid rgb(83, 83, 86);
 	width: 100%;
 	padding: 8px 10px;
 	border-radius: 2px;
 }
 .data .form-content form .names input:focus{
-	border: 3px solid #08762e;
+	border: 1px solid #08762e;
 	outline: none;
 }
 .data .form-content form .amount{
 	display: flex;
 	flex-direction: row;
 	align-items: center;
+	justify-content: space-between;
 	gap: 20px;
 	width: 100%;
 	padding-bottom: 15px;
 }
 .data .form-content form .amount .amountIpt{
-	width: 40%;
+	width: 20%;
 }
 .data .form-content form .amount .amountIpt input{
 	width: 100%;
@@ -407,7 +440,7 @@ table .res-head {
 	color: #9c9c9c;
 }
 .data .form-content form .amount div .amountIpt{
-	border: 3px solid rgb(83, 83, 86);
+	border: 1px solid rgb(83, 83, 86);
 	padding: 8px 10px;
 	border-radius: 2px;
 }
@@ -416,11 +449,18 @@ table .res-head {
 	outline: none;
 }
 .data .form-content form .amount .casual{
-	width: 50%;
+	width: 30%;
 	padding: 8px 10px;
 }
+.data .form-content form .amount .all-payTime {
+	width: 100%;
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: space-between;
+}
 .data .form-content form .amount .payTime{
-	width: 20%;
+	/* width: 20%; */
 }
 
 .data .form-content form .amount .payTime label{
@@ -428,7 +468,7 @@ table .res-head {
 	color: #9c9c9c;
 }
 .data .form-content form .amount .payTime input{
-	border: 3px solid #08762e;
+	border: 1px solid #08762e;
 	padding: 4px 35px;
 }
 .data .form-content form .amount .payTime input:hover{
@@ -530,8 +570,8 @@ table .res-head {
 	flex-direction: row;
 	align-items: center;
 	justify-content: center;
-	width: 15rem;
-	padding: 5px 0;
+	width: 10rem;
+	padding: 5px 5px;
 	border: .5px solid #08762e;
 	background-color: #fff;
 	box-shadow: 0 5px 10px rgba(0, 0, 0, 0.1);
@@ -576,7 +616,6 @@ table .res-head {
 	display: flex;
 	flex-direction: row;
 	align-items: flex-start;
-	/* justify-content: space-between; */
 	gap: 20px;
 	margin-bottom: 20px;
 }
